+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Layouts
 
+- Add `AnalyticsShell` component. [#307](https://github.com/mapbox/dr-ui/pull/307)
+
 ## Main
 
 - Add `ErrorBoundary` component. [#272](https://github.com/mapbox/dr-ui/pull/272)